webapp/requirements.txt
Flask
requests
python-dotenv
pymongo
flask_pymongo
sklearn
